Add better array/struct/union handling as fields/arguments
The logic in this module is becoming pretty convoluted, but essentially what's going on when transforming a cython type node to ctypes; if a type is categorized as a non-native type, I default it's ctype value to None, and it's string value to the passed in name. Setting it's ctype to None lets it run fine through ctypes-configure. After the configure I detect these types again and appropriately bubble up the orginally passed in name.

Essentially type checking has been offloaded to a run-time problem, and we now take no responsibility for verifying types passed into structs/unions/function calls as actually existing at compile time. We should add integration tests as a post-compile step to close this gap?